# website/src/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "elder-scrobz-api",
    "description": "",
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "paths": {
    "/api/v1/admin/scan": {
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Scan database",
        "description": "Scan database and fetch medatada",
        "operationId": "scan_db",
        "parameters": [
          {
            "name": "force",
            "in": "query",
            "description": "Scan all scrobbles instead of only unprocessed ones",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "coverart_only",
            "in": "query",
            "description": "Only fetch missing coverart",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Launch an asynchronous scrobble scan",
            "content": {
              "application/json": {}
            }
          }
        }
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "api_key": {
        "type": "apiKey",
        "in": "header",
        "name": "todo_apikey"
      }
    }
  },
...
}